move metrics code to models package
