+++ sherlock/internal/models/metrics_cron_test.go
@@ -1,0 +1,11 @@
+package models
+
+import (
+	"context"
+)
+
+func (s *modelSuite) TestUpdateMetrics() {
+	s.Run("doesn't error when running custom SQL", func() {
+		s.NoError(UpdateMetrics(context.Background(), s.DB))
+	})
+}
